bybit-botter/bybit_bot/requirements.txt
pandas
numpy
TA-Lib
websocket-client
bybit
python-dotenv
